Use the name of the desktop (Plasma) for KDE tasks
The name of the desktop is Plasma. Use the long name for now
(producer + product name, so KDE Plasma) in the description
of tasks which refers to the desktop product by KDE.
Even if the internal names of the tasks only refer to "kde",
the user-visibile strings now show the correct name.