Improve print service task and description
Using «print server» for the task name and description seems rather
confusing, depending on the user expectations and experience. This tends
to signal that this machine will be serving other systems, instead of
making it obvious that it is also required for local printing.

Change the name to print service which is more generic, and does not
have the same connotations as print server. Improve its description to
match. Add a Provides for backwards compatibility. Move the task from
server to user Section.

Closes: #696658
Ref: https://lwn.net/Articles/792952/