Respect tasksel/desktop debconf setting only if it's non empty
The default value is empty and thus we never show the default desktop
that we have selected. Turns out that the generic desktop tasks does
depend on gnome by default so we do get gnome anyway but we also get
it for ARM architectures where we wanted to install Xfce
(cf default_desktop file).