Fix the name of the Acquire::Retries option passed to apt-get
The option is not part of the "APT" group and the desired effect
was thus not achieved. Failures were instantly fatal without any retry.

This is particularly annoying when the failure is not due to the
network but due to bugs in APT, for example this one that I hit:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=973861